+++ NEWS.md
@@ -1,6 +1,24 @@
 # HYPEtools NEWS
 
-## HYPEtools 1.6.0
+## HYPEtools 1.6.1
+*Enhancements and bug fixes*
+
+#### Highlights
+* Add cbind option for adding groups in PlotPerformanceByAttribute()
+* Handle NA values when scaling axes in PlotPerformanceByAttribute()
+* Handle group IDs as strings in PlotPerformanceByAttribute()
+* Add argument to adjust spacing between plots and summary table in PlotPerformanceByAttribute()
+* Add warning for non-standard column names when using ReadGeoData()
+* Update default info columns for VariableLookup functions
+* Add jitter to points with duplicate geometries in PlotMapPoints()
+* Hide label group by default for Leaflet maps in MapRegionalSources()
+* Allow SUBIDs with many digits in SortGeoData()
+* Update PlotSimObsRegime() to ignore NA values and to fix percentile polygons when using log scale with negative/zero values (Issue #61)
+* Create distinctColorPalette() function and update PlotSubbasinRouting() and MapRegionalSources() to allow setting a seed so palettes can be reproduced
+* Update MapRegionalSources() to allow specifying different connection groups with different colors
+* Various updates to help text
+
+## HYPEtools 1.6.0 (2023-10-18)
 *Enhancements*
 
 #### Highlights